requirements.txt
Django==5.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
